wrote the ASOS data parser
next need to write the histogram generator for the wind speed and polar plot for the directions. 

I used object oriented programming to parse the data, so it should be easy to extend this data analysis to other projects in the future. for example, this script could easily access all the wind speeds for a certain time of day, if we want to model the expected wind for the time of day the lanch would happen. also, most of it's values are defined as constants, so it should be able to apply to other wind data sources with minor changes.